+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.36] - 2023-02-23
+## [0.0.37] - 2023-03-01
+### Added
+* static `Diagnostics` class for diagnostics functions like `TimeFunctionWithStopwatch`
+### Changed
+* directory for `LogManager` to Diagnostics
+
+## [0.0.36] - 2023-02-28
 ### Added
 * `Cast<TIn, TOut>()` to`ListExtensions` to convert a List from one type to another
 ### Removed